requirements.txt
altgraph==0.17.3
astroid==2.12.13
attrs==22.1.0
beautifulsoup4==4.11.1
bs4==0.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
eradicate==2.1.0
exceptiongroup==1.0.4
future==0.18.2
iniconfig==1.1.1
isort==5.11.2
lazy-object-proxy==1.8.0
lxml==4.9.2
mando==0.6.4
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
packaging==22.0
platformdirs==2.6.0
pluggy==1.0.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.1.1
pyflakes==3.0.1
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pylama==8.4.1
pylint==2.15.8
PyQt6==6.4.0
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
pytest==7.2.0
python-lzf==0.2.4
radon==5.1.0
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
vulture==2.6
wrapt==1.14.1
xmltodict==0.13.0
